+++ backend/canisters/community/impl/src/updates/add_reaction.rs
@@ -64,6 +64,7 @@ fn add_reaction_impl(args: Args, ext_caller: Option<Caller>, state: &mut Runtime
         args.message_id,
         args.reaction.clone(),
         now,
+        &state.data.migrated_user_ids,
         CommunityEventPusher {
             now,
             rng: state.env.rng(),